+++ packages/pigeon/platform_tests/alternate_language_test_plugin/android/src/main/java/com/example/alternate_language_test_plugin/AlternateLanguageTestPlugin.java
@@ -26,15 +26,19 @@ import java.util.List;
 import java.util.Map;
 
 /** This plugin handles the native side of the integration tests in example/integration_test/. */
-public class AlternateLanguageTestPlugin implements FlutterPlugin, HostIntegrationCoreApi {
+public class AlternateLanguageTestPlugin
+    implements FlutterPlugin, HostIntegrationCoreApi, CoreTests.HostCallbackCoreApi {
   @Nullable FlutterIntegrationCoreApi flutterApi = null;
+  @Nullable CoreTests.FlutterCallbackCoreApi flutterCallbackApi = null;
   @Nullable FlutterSmallApi flutterSmallApiOne = null;
   @Nullable FlutterSmallApi flutterSmallApiTwo = null;
 
   @Override
   public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
     HostIntegrationCoreApi.setUp(binding.getBinaryMessenger(), this);
+    CoreTests.HostCallbackCoreApi.setUp(binding.getBinaryMessenger(), this);
     flutterApi = new FlutterIntegrationCoreApi(binding.getBinaryMessenger());
+    flutterCallbackApi = new CoreTests.FlutterCallbackCoreApi(binding.getBinaryMessenger());
     flutterSmallApiOne = new FlutterSmallApi(binding.getBinaryMessenger(), "suffixOne");
     flutterSmallApiTwo = new FlutterSmallApi(binding.getBinaryMessenger(), "suffixTwo");
     TestPluginWithSuffix testSuffixApiOne = new TestPluginWithSuffix();
@@ -44,7 +48,46 @@ public class AlternateLanguageTestPlugin implements FlutterPlugin, HostIntegrati
   }
 
   @Override
-  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {}
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    HostIntegrationCoreApi.setUp(binding.getBinaryMessenger(), null);
+    CoreTests.HostCallbackCoreApi.setUp(binding.getBinaryMessenger(), null);
+  }
+
+  // HostCallbackCoreApi
+  @Override
+  public void noop(VoidResult result) {
+    result.success();
+  }
+
+  @Override
+  public void echoString(@NonNull String aString, Result<String> result) {
+    result.success(aString);
+  }
+
+  @Override
+  public void echoAllTypes(@NonNull AllTypes everything, Result<AllTypes> result) {
+    result.success(everything);
+  }
+
+  @Override
+  public void echoNullableString(@Nullable String aString, Result<String> result) {
+    result.success(aString);
+  }
+
+  @Override
+  public void throwError(Result<Object> result) {
+    result.error(new CoreTests.FlutterError("code", "message", "details"));
+  }
+
+  @Override
+  public void throwErrorFromVoid(VoidResult result) {
+    result.error(new CoreTests.FlutterError("code", "message", "details"));
+  }
+
+  @Override
+  public void taskQueueIsBackgroundThread(Result<Boolean> result) {
+    result.success(Looper.myLooper() != Looper.getMainLooper());
+  }
 
   // HostIntegrationCoreApi
 
@@ -1059,6 +1102,31 @@ public class AlternateLanguageTestPlugin implements FlutterPlugin, HostIntegrati
           }
         };
     flutterSmallApiOne.echoString(aString, resultCallbackOne);
+  }
+
+  @Override
+  public void callFlutterCallbackNoop(@NonNull VoidResult result) {
+    assert flutterCallbackApi != null;
+    flutterCallbackApi.noop(result);
+  }
+
+  @Override
+  public void callFlutterCallbackEchoString(
+      @NonNull String aString, @NonNull Result<String> result) {
+    assert flutterCallbackApi != null;
+    flutterCallbackApi.echoString(aString, result);
+  }
+
+  @Override
+  public void callFlutterCallbackThrowError(@NonNull NullableResult<Object> result) {
+    assert flutterCallbackApi != null;
+    flutterCallbackApi.throwError(result);
+  }
+
+  @Override
+  public void callFlutterCallbackThrowErrorFromVoid(@NonNull VoidResult result) {
+    assert flutterCallbackApi != null;
+    flutterCallbackApi.throwErrorFromVoid(result);
   }
 
   public @NonNull CoreTests.UnusedClass testIfUnusedClassIsGenerated() {